+++ docs/changelog.rst
@@ -66,6 +66,7 @@ Primary Modules
 Changes to the :doc:`primary modules <primary-modules>`.
 
 * Improved the error handling in the :py:func:`khoros.auth.get_session_key` function.
+* Merged the two ``if`` statements in the :py:func:`khoros.auth.get_sso_key` function.
 
 Supporting Modules
 ------------------